backend/requirements.txt
Django==4.2.2
Django-cors-headers==4.1.0
djangorestframework==3.14.0
pillow==10.0.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
psycopg==3.1.10
